split some helper functions into separate module
